Add brailleDictionary
Separando la logica del translator del diccionarioBraille.
