+++ documentation/apidocs/com/softtech/traductorbraille/GUI/class-use/BrailleCell.html
@@ -1,7 +1,7 @@
 <!DOCTYPE HTML>
 <html lang="es">
 <head>
-<!-- Generated by javadoc (17) on Fri Jul 05 22:29:43 ECT 2024 -->
+<!-- Generated by javadoc (17) on Fri Jul 05 23:46:13 ECT 2024 -->
 <title>Uses of Class com.softtech.traductorbraille.GUI.BrailleCell (TraductorBraille 1.0-SNAPSHOT API)</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
