+++ documentation/apidocs/com/softtech/traductorbraille/logic/class-use/BrailleDictionary.html
@@ -2,21 +2,21 @@
 <html lang="es">
 <head>
 <!-- Generated by javadoc (17) on Fri Jul 05 23:46:13 ECT 2024 -->
-<title>Uses of Class com.softtech.traductorbraille.logic.Printer.MirrorPrinter (TraductorBraille 1.0-SNAPSHOT API)</title>
+<title>Uses of Class com.softtech.traductorbraille.logic.BrailleDictionary (TraductorBraille 1.0-SNAPSHOT API)</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <meta name="dc.created" content="2024-07-05">
-<meta name="description" content="use: package: com.softtech.traductorbraille.logic.Printer, class: MirrorPrinter">
+<meta name="description" content="use: package: com.softtech.traductorbraille.logic, class: BrailleDictionary">
 <meta name="generator" content="javadoc/ClassUseWriter">
-<link rel="stylesheet" type="text/css" href="../../../../../../stylesheet.css" title="Style">
-<link rel="stylesheet" type="text/css" href="../../../../../../script-dir/jquery-ui.min.css" title="Style">
-<link rel="stylesheet" type="text/css" href="../../../../../../jquery-ui.overrides.css" title="Style">
-<script type="text/javascript" src="../../../../../../script.js"></script>
-<script type="text/javascript" src="../../../../../../script-dir/jquery-3.6.1.min.js"></script>
-<script type="text/javascript" src="../../../../../../script-dir/jquery-ui.min.js"></script>
+<link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
+<link rel="stylesheet" type="text/css" href="../../../../../script-dir/jquery-ui.min.css" title="Style">
+<link rel="stylesheet" type="text/css" href="../../../../../jquery-ui.overrides.css" title="Style">
+<script type="text/javascript" src="../../../../../script.js"></script>
+<script type="text/javascript" src="../../../../../script-dir/jquery-3.6.1.min.js"></script>
+<script type="text/javascript" src="../../../../../script-dir/jquery-ui.min.js"></script>
 </head>
 <body class="class-use-page">
-<script type="text/javascript">var pathtoroot = "../../../../../../";
+<script type="text/javascript">var pathtoroot = "../../../../../";
 loadScripts(document, 'script');</script>
 <noscript>
 <div>JavaScript is disabled on your browser.</div>
@@ -28,13 +28,13 @@ loadScripts(document, 'script');</script>
 <div class="top-nav" id="navbar-top">
 <div class="skip-nav"><a href="#skip-navbar-top" title="Skip navigation links">Skip navigation links</a></div>
 <ul id="navbar-top-firstrow" class="nav-list" title="Navigation">
-<li><a href="../../../../../../index.html">Overview</a></li>
+<li><a href="../../../../../index.html">Overview</a></li>
 <li><a href="../package-summary.html">Package</a></li>
-<li><a href="../MirrorPrinter.html" title="class in com.softtech.traductorbraille.logic.Printer">Class</a></li>
+<li><a href="../BrailleDictionary.html" title="class in com.softtech.traductorbraille.logic">Class</a></li>
 <li class="nav-bar-cell1-rev">Use</li>
 <li><a href="../package-tree.html">Tree</a></li>
-<li><a href="../../../../../../index-all.html">Index</a></li>
-<li><a href="../../../../../../help-doc.html#use">Help</a></li>
+<li><a href="../../../../../index-all.html">Index</a></li>
+<li><a href="../../../../../help-doc.html#use">Help</a></li>
 </ul>
 </div>
 <div class="sub-nav">
@@ -49,9 +49,9 @@ loadScripts(document, 'script');</script>
 <div class="flex-content">
 <main role="main">
 <div class="header">
-<h1 title="Uses of Class com.softtech.traductorbraille.logic.Printer.MirrorPrinter" class="title">Uses of Class<br>com.softtech.traductorbraille.logic.Printer.MirrorPrinter</h1>
+<h1 title="Uses of Class com.softtech.traductorbraille.logic.BrailleDictionary" class="title">Uses of Class<br>com.softtech.traductorbraille.logic.BrailleDictionary</h1>
 </div>
-No usage of com.softtech.traductorbraille.logic.Printer.MirrorPrinter</main>
+No usage of com.softtech.traductorbraille.logic.BrailleDictionary</main>
 <footer role="contentinfo">
 <hr>
 <p class="legal-copy"><small>Copyright &#169; 2024. All rights reserved.</small></p>
