+++ documentation/apidocs/com/softtech/traductorbraille/logic/package-summary.html
@@ -1,7 +1,7 @@
 <!DOCTYPE HTML>
 <html lang="es">
 <head>
-<!-- Generated by javadoc (17) on Fri Jul 05 22:29:43 ECT 2024 -->
+<!-- Generated by javadoc (17) on Fri Jul 05 23:46:13 ECT 2024 -->
 <title>com.softtech.traductorbraille.logic (TraductorBraille 1.0-SNAPSHOT API)</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
@@ -80,12 +80,14 @@ loadScripts(document, 'script');</script>
 <div class="summary-table two-column-summary">
 <div class="table-header col-first">Class</div>
 <div class="table-header col-last">Description</div>
-<div class="col-first even-row-color class-summary class-summary-tab2"><a href="ExportService.html" title="class in com.softtech.traductorbraille.logic">ExportService</a></div>
-<div class="col-last even-row-color class-summary class-summary-tab2">
+<div class="col-first even-row-color class-summary class-summary-tab2"><a href="BrailleDictionary.html" title="class in com.softtech.traductorbraille.logic">BrailleDictionary</a></div>
+<div class="col-last even-row-color class-summary class-summary-tab2">&nbsp;</div>
+<div class="col-first odd-row-color class-summary class-summary-tab2"><a href="ExportService.html" title="class in com.softtech.traductorbraille.logic">ExportService</a></div>
+<div class="col-last odd-row-color class-summary class-summary-tab2">
 <div class="block">Servicio de exportación para convertir texto en Braille a diferentes formatos como TXT, PDF e imágenes.</div>
 </div>
-<div class="col-first odd-row-color class-summary class-summary-tab2"><a href="Translator.html" title="class in com.softtech.traductorbraille.logic">Translator</a></div>
-<div class="col-last odd-row-color class-summary class-summary-tab2">
+<div class="col-first even-row-color class-summary class-summary-tab2"><a href="Translator.html" title="class in com.softtech.traductorbraille.logic">Translator</a></div>
+<div class="col-last even-row-color class-summary class-summary-tab2">
 <div class="block">Translator aloja la lógica de traducción y diccionario de equivalencias entre el abecedario español y las claves braille.</div>
 </div>
 </div>
