+++ documentation/apidocs/com/softtech/traductorbraille/logic/package-tree.html
@@ -1,7 +1,7 @@
 <!DOCTYPE HTML>
 <html lang="es">
 <head>
-<!-- Generated by javadoc (17) on Fri Jul 05 22:29:43 ECT 2024 -->
+<!-- Generated by javadoc (17) on Fri Jul 05 23:46:13 ECT 2024 -->
 <title>com.softtech.traductorbraille.logic Class Hierarchy (TraductorBraille 1.0-SNAPSHOT API)</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
@@ -60,6 +60,7 @@ loadScripts(document, 'script');</script>
 <ul>
 <li class="circle">java.lang.<a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Object.html" class="type-name-link external-link" title="class or interface in java.lang">Object</a>
 <ul>
+<li class="circle">com.softtech.traductorbraille.logic.<a href="BrailleDictionary.html" class="type-name-link" title="class in com.softtech.traductorbraille.logic">BrailleDictionary</a></li>
 <li class="circle">com.softtech.traductorbraille.logic.<a href="ExportService.html" class="type-name-link" title="class in com.softtech.traductorbraille.logic">ExportService</a></li>
 <li class="circle">com.softtech.traductorbraille.logic.<a href="Translator.html" class="type-name-link" title="class in com.softtech.traductorbraille.logic">Translator</a></li>
 </ul>
