+++ documentation/apidocs/constant-values.html
@@ -2,12 +2,12 @@
 <html lang="es">
 <head>
 <!-- Generated by javadoc (17) on Fri Jul 05 23:46:13 ECT 2024 -->
-<title>All Packages (TraductorBraille 1.0-SNAPSHOT API)</title>
+<title>Constant Field Values (TraductorBraille 1.0-SNAPSHOT API)</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <meta name="dc.created" content="2024-07-05">
-<meta name="description" content="package index">
-<meta name="generator" content="javadoc/AllPackagesIndexWriter">
+<meta name="description" content="summary of constants">
+<meta name="generator" content="javadoc/ConstantsSummaryWriterImpl">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
 <link rel="stylesheet" type="text/css" href="script-dir/jquery-ui.min.css" title="Style">
 <link rel="stylesheet" type="text/css" href="jquery-ui.overrides.css" title="Style">
@@ -15,7 +15,7 @@
 <script type="text/javascript" src="script-dir/jquery-3.6.1.min.js"></script>
 <script type="text/javascript" src="script-dir/jquery-ui.min.js"></script>
 </head>
-<body class="all-packages-index-page">
+<body class="constants-summary-page">
 <script type="text/javascript">var pathtoroot = "./";
 loadScripts(document, 'script');</script>
 <noscript>
@@ -34,7 +34,7 @@ loadScripts(document, 'script');</script>
 <li>Use</li>
 <li><a href="overview-tree.html">Tree</a></li>
 <li><a href="index-all.html">Index</a></li>
-<li><a href="help-doc.html#all-packages">Help</a></li>
+<li><a href="help-doc.html#constant-values">Help</a></li>
 </ul>
 </div>
 <div class="sub-nav">
@@ -49,21 +49,33 @@ loadScripts(document, 'script');</script>
 <div class="flex-content">
 <main role="main">
 <div class="header">
-<h1 title="All&amp;nbsp;Packages" class="title">All&nbsp;Packages</h1>
+<h1 title="Constant Field Values" class="title">Constant Field Values</h1>
+<section class="packages">
+<h2 title="Contents">Contents</h2>
+<ul>
+<li><a href="#com.softtech">com.softtech.*</a></li>
+</ul>
+</section>
 </div>
-<div class="caption"><span>Package Summary</span></div>
-<div class="summary-table two-column-summary">
-<div class="table-header col-first">Package</div>
-<div class="table-header col-last">Description</div>
-<div class="col-first even-row-color"><a href="com/softtech/traductorbraille/executable/package-summary.html">com.softtech.traductorbraille.executable</a></div>
-<div class="col-last even-row-color">&nbsp;</div>
-<div class="col-first odd-row-color"><a href="com/softtech/traductorbraille/GUI/package-summary.html">com.softtech.traductorbraille.GUI</a></div>
-<div class="col-last odd-row-color">&nbsp;</div>
-<div class="col-first even-row-color"><a href="com/softtech/traductorbraille/logic/package-summary.html">com.softtech.traductorbraille.logic</a></div>
-<div class="col-last even-row-color">&nbsp;</div>
-<div class="col-first odd-row-color"><a href="com/softtech/traductorbraille/logic/Printer/package-summary.html">com.softtech.traductorbraille.logic.Printer</a></div>
-<div class="col-last odd-row-color">&nbsp;</div>
+<section class="constants-summary" id="com.softtech.traductorbraille.logic.Printer">
+<h2 title="com.softtech">com.softtech.*</h2>
+<ul class="block-list">
+<li>
+<div class="caption"><span>com.softtech.traductorbraille.logic.Printer.<a href="com/softtech/traductorbraille/logic/Printer/Printer.html" title="interface in com.softtech.traductorbraille.logic.Printer">Printer</a></span></div>
+<div class="summary-table three-column-summary">
+<div class="table-header col-first">Modifier and Type</div>
+<div class="table-header col-second">Constant Field</div>
+<div class="table-header col-last">Value</div>
+<div class="col-first even-row-color"><code id="com.softtech.traductorbraille.logic.Printer.Printer.PDF_HEIGTH">public&nbsp;static&nbsp;final&nbsp;int</code></div>
+<div class="col-second even-row-color"><code><a href="com/softtech/traductorbraille/logic/Printer/Printer.html#PDF_HEIGTH">PDF_HEIGTH</a></code></div>
+<div class="col-last even-row-color"><code>1056</code></div>
+<div class="col-first odd-row-color"><code id="com.softtech.traductorbraille.logic.Printer.Printer.PDF_WIDTH">public&nbsp;static&nbsp;final&nbsp;int</code></div>
+<div class="col-second odd-row-color"><code><a href="com/softtech/traductorbraille/logic/Printer/Printer.html#PDF_WIDTH">PDF_WIDTH</a></code></div>
+<div class="col-last odd-row-color"><code>816</code></div>
 </div>
+</li>
+</ul>
+</section>
 </main>
 <footer role="contentinfo">
 <hr>
