+++ documentation/apidocs/help-doc.html
@@ -1,7 +1,7 @@
 <!DOCTYPE HTML>
 <html lang="es">
 <head>
-<!-- Generated by javadoc (17) on Fri Jul 05 22:29:43 ECT 2024 -->
+<!-- Generated by javadoc (17) on Fri Jul 05 23:46:13 ECT 2024 -->
 <title>API Help (TraductorBraille 1.0-SNAPSHOT API)</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
@@ -70,6 +70,7 @@ loadScripts(document, 'script');</script>
 <li><a href="#doc-file">Other Files</a></li>
 <li><a href="#use">Use</a></li>
 <li><a href="#tree">Tree (Class Hierarchy)</a></li>
+<li><a href="#constant-values">Constant Field Values</a></li>
 <li><a href="#serialized-form">Serialized Form</a></li>
 <li><a href="#all-packages">All Packages</a></li>
 <li><a href="#all-classes">All Classes and Interfaces</a></li>
@@ -161,6 +162,10 @@ The following sections describe the different kinds of pages in this collection.
 <li>When viewing the Overview page, clicking on TREE displays the hierarchy for all packages.</li>
 <li>When viewing a particular package, class or interface page, clicking on TREE displays the hierarchy for only that package.</li>
 </ul>
+</section>
+<section class="help-section" id="constant-values">
+<h3>Constant Field Values</h3>
+<p>The <a href="constant-values.html">Constant Field Values</a> page lists the static final fields and their values.</p>
 </section>
 <section class="help-section" id="serialized-form">
 <h3>Serialized Form</h3>
