+++ documentation/apidocs/index-all.html
@@ -1,7 +1,7 @@
 <!DOCTYPE HTML>
 <html lang="es">
 <head>
-<!-- Generated by javadoc (17) on Fri Jul 05 22:29:43 ECT 2024 -->
+<!-- Generated by javadoc (17) on Fri Jul 05 23:46:13 ECT 2024 -->
 <title>Index (TraductorBraille 1.0-SNAPSHOT API)</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
@@ -51,7 +51,7 @@ loadScripts(document, 'script');</script>
 <div class="header">
 <h1>Index</h1>
 </div>
-<a href="#I:B">B</a>&nbsp;<a href="#I:C">C</a>&nbsp;<a href="#I:D">D</a>&nbsp;<a href="#I:E">E</a>&nbsp;<a href="#I:G">G</a>&nbsp;<a href="#I:J">J</a>&nbsp;<a href="#I:M">M</a>&nbsp;<a href="#I:N">N</a>&nbsp;<a href="#I:P">P</a>&nbsp;<a href="#I:S">S</a>&nbsp;<a href="#I:T">T</a>&nbsp;<br><a href="allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="serialized-form.html">Serialized&nbsp;Form</a>
+<a href="#I:B">B</a>&nbsp;<a href="#I:C">C</a>&nbsp;<a href="#I:D">D</a>&nbsp;<a href="#I:E">E</a>&nbsp;<a href="#I:G">G</a>&nbsp;<a href="#I:J">J</a>&nbsp;<a href="#I:M">M</a>&nbsp;<a href="#I:N">N</a>&nbsp;<a href="#I:P">P</a>&nbsp;<a href="#I:S">S</a>&nbsp;<a href="#I:T">T</a>&nbsp;<br><a href="allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="constant-values.html">Constant&nbsp;Field&nbsp;Values</a><span class="vertical-separator">|</span><a href="serialized-form.html">Serialized&nbsp;Form</a>
 <h2 class="title" id="I:B">B</h2>
 <dl class="index">
 <dt><a href="com/softtech/traductorbraille/GUI/BrailleCell.html" class="type-name-link" title="class in com.softtech.traductorbraille.GUI">BrailleCell</a> - Class in <a href="com/softtech/traductorbraille/GUI/package-summary.html">com.softtech.traductorbraille.GUI</a></dt>
@@ -62,6 +62,10 @@ loadScripts(document, 'script');</script>
 <dd>
 <div class="block">Constructor de la clase BrailleCell.</div>
 </dd>
+<dt><a href="com/softtech/traductorbraille/logic/BrailleDictionary.html" class="type-name-link" title="class in com.softtech.traductorbraille.logic">BrailleDictionary</a> - Class in <a href="com/softtech/traductorbraille/logic/package-summary.html">com.softtech.traductorbraille.logic</a></dt>
+<dd>&nbsp;</dd>
+<dt><a href="com/softtech/traductorbraille/logic/BrailleDictionary.html#%3Cinit%3E()" class="member-name-link">BrailleDictionary()</a> - Constructor for class com.softtech.traductorbraille.logic.<a href="com/softtech/traductorbraille/logic/BrailleDictionary.html" title="class in com.softtech.traductorbraille.logic">BrailleDictionary</a></dt>
+<dd>&nbsp;</dd>
 <dt><a href="com/softtech/traductorbraille/logic/Translator.html#brailleToUnicode(java.lang.String)" class="member-name-link">brailleToUnicode(String)</a> - Method in class com.softtech.traductorbraille.logic.<a href="com/softtech/traductorbraille/logic/Translator.html" title="class in com.softtech.traductorbraille.logic">Translator</a></dt>
 <dd>
 <div class="block">convertir texto braille a cuadratin Unicode</div>
@@ -134,7 +138,7 @@ loadScripts(document, 'script');</script>
 </dl>
 <h2 class="title" id="I:G">G</h2>
 <dl class="index">
-<dt><a href="com/softtech/traductorbraille/logic/Translator.html#generateBrailleMirror(java.lang.String)" class="member-name-link">generateBrailleMirror(String)</a> - Method in class com.softtech.traductorbraille.logic.<a href="com/softtech/traductorbraille/logic/Translator.html" title="class in com.softtech.traductorbraille.logic">Translator</a></dt>
+<dt><a href="com/softtech/traductorbraille/logic/BrailleDictionary.html#generateBrailleMirror(java.lang.String)" class="member-name-link">generateBrailleMirror(String)</a> - Method in class com.softtech.traductorbraille.logic.<a href="com/softtech/traductorbraille/logic/BrailleDictionary.html" title="class in com.softtech.traductorbraille.logic">BrailleDictionary</a></dt>
 <dd>
 <div class="block">transformar texto braille en su forma expejo</div>
 </dd>
@@ -228,6 +232,10 @@ loadScripts(document, 'script');</script>
 <dd>
 <div class="block">Método para pintar los puntos Braille en la celda.</div>
 </dd>
+<dt><a href="com/softtech/traductorbraille/logic/Printer/Printer.html#PDF_HEIGTH" class="member-name-link">PDF_HEIGTH</a> - Static variable in interface com.softtech.traductorbraille.logic.Printer.<a href="com/softtech/traductorbraille/logic/Printer/Printer.html" title="interface in com.softtech.traductorbraille.logic.Printer">Printer</a></dt>
+<dd>&nbsp;</dd>
+<dt><a href="com/softtech/traductorbraille/logic/Printer/Printer.html#PDF_WIDTH" class="member-name-link">PDF_WIDTH</a> - Static variable in interface com.softtech.traductorbraille.logic.Printer.<a href="com/softtech/traductorbraille/logic/Printer/Printer.html" title="interface in com.softtech.traductorbraille.logic.Printer">Printer</a></dt>
+<dd>&nbsp;</dd>
 <dt><a href="com/softtech/traductorbraille/logic/Printer/MirrorPrinter.html#print(java.lang.String,int)" class="member-name-link">print(String, int)</a> - Method in class com.softtech.traductorbraille.logic.Printer.<a href="com/softtech/traductorbraille/logic/Printer/MirrorPrinter.html" title="class in com.softtech.traductorbraille.logic.Printer">MirrorPrinter</a></dt>
 <dd>
 <div class="block">Imprime el texto dado en el tamaño de fuente especificado.</div>
@@ -271,7 +279,7 @@ loadScripts(document, 'script');</script>
 <dt><a href="com/softtech/traductorbraille/logic/Translator.html#%3Cinit%3E()" class="member-name-link">Translator()</a> - Constructor for class com.softtech.traductorbraille.logic.<a href="com/softtech/traductorbraille/logic/Translator.html" title="class in com.softtech.traductorbraille.logic">Translator</a></dt>
 <dd>&nbsp;</dd>
 </dl>
-<a href="#I:B">B</a>&nbsp;<a href="#I:C">C</a>&nbsp;<a href="#I:D">D</a>&nbsp;<a href="#I:E">E</a>&nbsp;<a href="#I:G">G</a>&nbsp;<a href="#I:J">J</a>&nbsp;<a href="#I:M">M</a>&nbsp;<a href="#I:N">N</a>&nbsp;<a href="#I:P">P</a>&nbsp;<a href="#I:S">S</a>&nbsp;<a href="#I:T">T</a>&nbsp;<br><a href="allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="serialized-form.html">Serialized&nbsp;Form</a></main>
+<a href="#I:B">B</a>&nbsp;<a href="#I:C">C</a>&nbsp;<a href="#I:D">D</a>&nbsp;<a href="#I:E">E</a>&nbsp;<a href="#I:G">G</a>&nbsp;<a href="#I:J">J</a>&nbsp;<a href="#I:M">M</a>&nbsp;<a href="#I:N">N</a>&nbsp;<a href="#I:P">P</a>&nbsp;<a href="#I:S">S</a>&nbsp;<a href="#I:T">T</a>&nbsp;<br><a href="allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="allpackages-index.html">All&nbsp;Packages</a><span class="vertical-separator">|</span><a href="constant-values.html">Constant&nbsp;Field&nbsp;Values</a><span class="vertical-separator">|</span><a href="serialized-form.html">Serialized&nbsp;Form</a></main>
 <footer role="contentinfo">
 <hr>
 <p class="legal-copy"><small>Copyright &#169; 2024. All rights reserved.</small></p>
