+++ documentation/apidocs/member-search-index.js
@@ -1,1 +1,1 @@
-memberSearchIndex = [{"p":"com.softtech.traductorbraille.GUI","c":"BrailleCell","l":"BrailleCell()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.logic","c":"Translator","l":"brailleToUnicode(String)","u":"brailleToUnicode(java.lang.String)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"MirrorPrinter","l":"calculateLines(Graphics2D, String, PageFormat, int)","u":"calculateLines(java.awt.Graphics2D,java.lang.String,java.awt.print.PageFormat,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"NormalPrinter","l":"calculateLines(Graphics2D, String, PageFormat, int)","u":"calculateLines(java.awt.Graphics2D,java.lang.String,java.awt.print.PageFormat,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"Printer","l":"calculateLines(Graphics2D, String, PageFormat, int)","u":"calculateLines(java.awt.Graphics2D,java.lang.String,java.awt.print.PageFormat,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"MirrorPrinter","l":"calculateLinesForParagraph(Graphics2D, String[], PageFormat)","u":"calculateLinesForParagraph(java.awt.Graphics2D,java.lang.String[],java.awt.print.PageFormat)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"NormalPrinter","l":"calculateLinesForParagraph(Graphics2D, String[], PageFormat)","u":"calculateLinesForParagraph(java.awt.Graphics2D,java.lang.String[],java.awt.print.PageFormat)"},{"p":"com.softtech.traductorbraille.GUI","c":"BrailleCell","l":"clearPoints()"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"Printer","l":"createContentPanel(String, int)","u":"createContentPanel(java.lang.String,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"MirrorPrinter","l":"drawText(Graphics2D, List<String>, PageFormat, int)","u":"drawText(java.awt.Graphics2D,java.util.List,java.awt.print.PageFormat,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"NormalPrinter","l":"drawText(Graphics2D, List<String>, PageFormat, int)","u":"drawText(java.awt.Graphics2D,java.util.List,java.awt.print.PageFormat,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"Printer","l":"drawText(Graphics2D, List<String>, PageFormat, int)","u":"drawText(java.awt.Graphics2D,java.util.List,java.awt.print.PageFormat,int)"},{"p":"com.softtech.traductorbraille.logic","c":"ExportService","l":"exportBraille(File, String, String, int, boolean, boolean, Color)","u":"exportBraille(java.io.File,java.lang.String,java.lang.String,int,boolean,boolean,java.awt.Color)"},{"p":"com.softtech.traductorbraille.logic","c":"ExportService","l":"ExportService()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.logic","c":"Translator","l":"generateBrailleMirror(String)","u":"generateBrailleMirror(java.lang.String)"},{"p":"com.softtech.traductorbraille.GUI","c":"BrailleCell","l":"getPoint(int)"},{"p":"com.softtech.traductorbraille.GUI","c":"JFExport","l":"getTexto()"},{"p":"com.softtech.traductorbraille.GUI","c":"JFExport","l":"JFExport(String)","u":"%3Cinit%3E(java.lang.String)"},{"p":"com.softtech.traductorbraille.GUI","c":"JFFileChooser","l":"JFFileChooser()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.GUI","c":"JFPreview","l":"JFPreview(String, int)","u":"%3Cinit%3E(java.lang.String,int)"},{"p":"com.softtech.traductorbraille.GUI","c":"JFTranslator","l":"JFTranslator()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.GUI","c":"JFTranslatorGUI","l":"JFTranslatorGUI()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.executable","c":"MainClass","l":"main(String[])","u":"main(java.lang.String[])"},{"p":"com.softtech.traductorbraille.executable","c":"MainClass","l":"MainClass()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"MirrorPrinter","l":"MirrorPrinter()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"NormalPrinter","l":"NormalPrinter()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.GUI","c":"BrailleCell","l":"paintComponent(Graphics)","u":"paintComponent(java.awt.Graphics)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"MirrorPrinter","l":"print(String, int)","u":"print(java.lang.String,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"NormalPrinter","l":"print(String, int)","u":"print(java.lang.String,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"Printer","l":"print(String, int)","u":"print(java.lang.String,int)"},{"p":"com.softtech.traductorbraille.GUI","c":"BrailleCell","l":"setPoint(int, boolean)","u":"setPoint(int,boolean)"},{"p":"com.softtech.traductorbraille.logic","c":"Translator","l":"translateToBraille(String)","u":"translateToBraille(java.lang.String)"},{"p":"com.softtech.traductorbraille.logic","c":"Translator","l":"translateToSpanish(String)","u":"translateToSpanish(java.lang.String)"},{"p":"com.softtech.traductorbraille.logic","c":"Translator","l":"translateToSpanishNumbersOnly(String)","u":"translateToSpanishNumbersOnly(java.lang.String)"},{"p":"com.softtech.traductorbraille.logic","c":"Translator","l":"Translator()","u":"%3Cinit%3E()"}];updateSearchResults();
+memberSearchIndex = [{"p":"com.softtech.traductorbraille.GUI","c":"BrailleCell","l":"BrailleCell()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.logic","c":"BrailleDictionary","l":"BrailleDictionary()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.logic","c":"Translator","l":"brailleToUnicode(String)","u":"brailleToUnicode(java.lang.String)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"MirrorPrinter","l":"calculateLines(Graphics2D, String, PageFormat, int)","u":"calculateLines(java.awt.Graphics2D,java.lang.String,java.awt.print.PageFormat,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"NormalPrinter","l":"calculateLines(Graphics2D, String, PageFormat, int)","u":"calculateLines(java.awt.Graphics2D,java.lang.String,java.awt.print.PageFormat,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"Printer","l":"calculateLines(Graphics2D, String, PageFormat, int)","u":"calculateLines(java.awt.Graphics2D,java.lang.String,java.awt.print.PageFormat,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"MirrorPrinter","l":"calculateLinesForParagraph(Graphics2D, String[], PageFormat)","u":"calculateLinesForParagraph(java.awt.Graphics2D,java.lang.String[],java.awt.print.PageFormat)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"NormalPrinter","l":"calculateLinesForParagraph(Graphics2D, String[], PageFormat)","u":"calculateLinesForParagraph(java.awt.Graphics2D,java.lang.String[],java.awt.print.PageFormat)"},{"p":"com.softtech.traductorbraille.GUI","c":"BrailleCell","l":"clearPoints()"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"Printer","l":"createContentPanel(String, int)","u":"createContentPanel(java.lang.String,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"MirrorPrinter","l":"drawText(Graphics2D, List<String>, PageFormat, int)","u":"drawText(java.awt.Graphics2D,java.util.List,java.awt.print.PageFormat,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"NormalPrinter","l":"drawText(Graphics2D, List<String>, PageFormat, int)","u":"drawText(java.awt.Graphics2D,java.util.List,java.awt.print.PageFormat,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"Printer","l":"drawText(Graphics2D, List<String>, PageFormat, int)","u":"drawText(java.awt.Graphics2D,java.util.List,java.awt.print.PageFormat,int)"},{"p":"com.softtech.traductorbraille.logic","c":"ExportService","l":"exportBraille(File, String, String, int, boolean, boolean, Color)","u":"exportBraille(java.io.File,java.lang.String,java.lang.String,int,boolean,boolean,java.awt.Color)"},{"p":"com.softtech.traductorbraille.logic","c":"ExportService","l":"ExportService()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.logic","c":"BrailleDictionary","l":"generateBrailleMirror(String)","u":"generateBrailleMirror(java.lang.String)"},{"p":"com.softtech.traductorbraille.GUI","c":"BrailleCell","l":"getPoint(int)"},{"p":"com.softtech.traductorbraille.GUI","c":"JFExport","l":"getTexto()"},{"p":"com.softtech.traductorbraille.GUI","c":"JFExport","l":"JFExport(String)","u":"%3Cinit%3E(java.lang.String)"},{"p":"com.softtech.traductorbraille.GUI","c":"JFFileChooser","l":"JFFileChooser()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.GUI","c":"JFPreview","l":"JFPreview(String, int)","u":"%3Cinit%3E(java.lang.String,int)"},{"p":"com.softtech.traductorbraille.GUI","c":"JFTranslator","l":"JFTranslator()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.GUI","c":"JFTranslatorGUI","l":"JFTranslatorGUI()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.executable","c":"MainClass","l":"main(String[])","u":"main(java.lang.String[])"},{"p":"com.softtech.traductorbraille.executable","c":"MainClass","l":"MainClass()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"MirrorPrinter","l":"MirrorPrinter()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"NormalPrinter","l":"NormalPrinter()","u":"%3Cinit%3E()"},{"p":"com.softtech.traductorbraille.GUI","c":"BrailleCell","l":"paintComponent(Graphics)","u":"paintComponent(java.awt.Graphics)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"Printer","l":"PDF_HEIGTH"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"Printer","l":"PDF_WIDTH"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"MirrorPrinter","l":"print(String, int)","u":"print(java.lang.String,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"NormalPrinter","l":"print(String, int)","u":"print(java.lang.String,int)"},{"p":"com.softtech.traductorbraille.logic.Printer","c":"Printer","l":"print(String, int)","u":"print(java.lang.String,int)"},{"p":"com.softtech.traductorbraille.GUI","c":"BrailleCell","l":"setPoint(int, boolean)","u":"setPoint(int,boolean)"},{"p":"com.softtech.traductorbraille.logic","c":"Translator","l":"translateToBraille(String)","u":"translateToBraille(java.lang.String)"},{"p":"com.softtech.traductorbraille.logic","c":"Translator","l":"translateToSpanish(String)","u":"translateToSpanish(java.lang.String)"},{"p":"com.softtech.traductorbraille.logic","c":"Translator","l":"translateToSpanishNumbersOnly(String)","u":"translateToSpanishNumbersOnly(java.lang.String)"},{"p":"com.softtech.traductorbraille.logic","c":"Translator","l":"Translator()","u":"%3Cinit%3E()"}];updateSearchResults();
