+++ documentation/apidocs/overview-tree.html
@@ -1,7 +1,7 @@
 <!DOCTYPE HTML>
 <html lang="es">
 <head>
-<!-- Generated by javadoc (17) on Fri Jul 05 22:29:43 ECT 2024 -->
+<!-- Generated by javadoc (17) on Fri Jul 05 23:46:13 ECT 2024 -->
 <title>Class Hierarchy (TraductorBraille 1.0-SNAPSHOT API)</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
@@ -63,6 +63,7 @@ loadScripts(document, 'script');</script>
 <ul>
 <li class="circle">java.lang.<a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Object.html" class="type-name-link external-link" title="class or interface in java.lang">Object</a>
 <ul>
+<li class="circle">com.softtech.traductorbraille.logic.<a href="com/softtech/traductorbraille/logic/BrailleDictionary.html" class="type-name-link" title="class in com.softtech.traductorbraille.logic">BrailleDictionary</a></li>
 <li class="circle">java.awt.<a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.desktop/java/awt/Component.html" class="type-name-link external-link" title="class or interface in java.awt">Component</a> (implements java.awt.image.<a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.desktop/java/awt/image/ImageObserver.html" title="class or interface in java.awt.image" class="external-link">ImageObserver</a>, java.awt.<a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.desktop/java/awt/MenuContainer.html" title="class or interface in java.awt" class="external-link">MenuContainer</a>, java.io.<a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/io/Serializable.html" title="class or interface in java.io" class="external-link">Serializable</a>)
 <ul>
 <li class="circle">java.awt.<a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.desktop/java/awt/Container.html" class="type-name-link external-link" title="class or interface in java.awt">Container</a>
