+++ documentation/apidocs/serialized-form.html
@@ -1,7 +1,7 @@
 <!DOCTYPE HTML>
 <html lang="es">
 <head>
-<!-- Generated by javadoc (17) on Fri Jul 05 22:29:43 ECT 2024 -->
+<!-- Generated by javadoc (17) on Fri Jul 05 23:46:13 ECT 2024 -->
 <title>Serialized Form (TraductorBraille 1.0-SNAPSHOT API)</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
@@ -237,6 +237,10 @@ loadScripts(document, 'script');</script>
 <li class="block-list">
 <h5>mirror</h5>
 <pre><a href="com/softtech/traductorbraille/logic/Printer/MirrorPrinter.html" title="class in com.softtech.traductorbraille.logic.Printer">MirrorPrinter</a> mirror</pre>
+</li>
+<li class="block-list">
+<h5>mirrorBraille</h5>
+<pre><a href="com/softtech/traductorbraille/logic/BrailleDictionary.html" title="class in com.softtech.traductorbraille.logic">BrailleDictionary</a> mirrorBraille</pre>
 </li>
 <li class="block-list">
 <h5>mirrorText</h5>
