+++ documentation/apidocs/type-search-index.js
@@ -1,1 +1,1 @@
-typeSearchIndex = [{"l":"All Classes and Interfaces","u":"allclasses-index.html"},{"p":"com.softtech.traductorbraille.GUI","l":"BrailleCell"},{"p":"com.softtech.traductorbraille.logic","l":"ExportService"},{"p":"com.softtech.traductorbraille.GUI","l":"JFExport"},{"p":"com.softtech.traductorbraille.GUI","l":"JFFileChooser"},{"p":"com.softtech.traductorbraille.GUI","l":"JFPreview"},{"p":"com.softtech.traductorbraille.GUI","l":"JFTranslator"},{"p":"com.softtech.traductorbraille.GUI","l":"JFTranslatorGUI"},{"p":"com.softtech.traductorbraille.executable","l":"MainClass"},{"p":"com.softtech.traductorbraille.logic.Printer","l":"MirrorPrinter"},{"p":"com.softtech.traductorbraille.logic.Printer","l":"NormalPrinter"},{"p":"com.softtech.traductorbraille.logic.Printer","l":"Printer"},{"p":"com.softtech.traductorbraille.logic","l":"Translator"}];updateSearchResults();
+typeSearchIndex = [{"l":"All Classes and Interfaces","u":"allclasses-index.html"},{"p":"com.softtech.traductorbraille.GUI","l":"BrailleCell"},{"p":"com.softtech.traductorbraille.logic","l":"BrailleDictionary"},{"p":"com.softtech.traductorbraille.logic","l":"ExportService"},{"p":"com.softtech.traductorbraille.GUI","l":"JFExport"},{"p":"com.softtech.traductorbraille.GUI","l":"JFFileChooser"},{"p":"com.softtech.traductorbraille.GUI","l":"JFPreview"},{"p":"com.softtech.traductorbraille.GUI","l":"JFTranslator"},{"p":"com.softtech.traductorbraille.GUI","l":"JFTranslatorGUI"},{"p":"com.softtech.traductorbraille.executable","l":"MainClass"},{"p":"com.softtech.traductorbraille.logic.Printer","l":"MirrorPrinter"},{"p":"com.softtech.traductorbraille.logic.Printer","l":"NormalPrinter"},{"p":"com.softtech.traductorbraille.logic.Printer","l":"Printer"},{"p":"com.softtech.traductorbraille.logic","l":"Translator"}];updateSearchResults();
